refactor parser tests
